requirements.txt
flask
transformers
torch
datasets
Pillow
nltk
evaluate

